services/payment-service/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.1
stripe==8.5.0
mysql-connector-python==8.3.0